Update on AppIcon, overall Data structure, and behavior of done Clicked
App Icon is set.
CourseVO's time is changed to DateInterval from just simple string.
When done button is clicked and If there is overlapped courses from the cart, alert shows up to restrict having overlapped course.

TODO for tomorrow:
When done button is clicked, the courses from the cart should be added to schedule.

To be considered from its behavior:
1. Is it going to be appended from existed course or whenever user schedule, it will going to reset?